refactor: split catalog to smaller files (#19870)
